% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_to_ret.R
\name{adapt_to_ret}
\alias{adapt_to_ret}
\title{Auxiliary function for enabling correct conversion from R dataframes to python and GAMS}
\usage{
adapt_to_ret(set_par)
}
\arguments{
\item{set_par}{dataframe corresponding to a set or a parameter to be loaded in ixmp}
}
\description{
Auxiliary function for enabling correct conversion from R dataframes to python and GAMS
}
\examples{
a.df = data.frame( i = c('a','b'), value = c(350 , 600) , unit = 'cases')
scen$add_par("a", adapt_to_ret(a.df))

}
