% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_rixmp.R
\name{rixmp}
\alias{rixmp}
\title{ixmp core}
\description{
This page provides documentation for the ixmp-R-reticulate implementation.
The package is sourcing the ixmp-core Python package, translating it with the R 'reticulate' package.
}
\details{
Main features:

 - a platform core linked to a database instance,

 - integration with the |MESSAGEix| Integrated Assessment model
   and other numerical models

 - for full documentation of the Python functions, use py_help(<package>$<class>)
}
